separate chartcustomization logic
